Moved models outside of folder
