+++ api/src/models/index.ts
@@ -1,6 +1,7 @@
-export * from './users/fsae-user.model';
-export * from './users/roles';
-export * from './users/member.model';
-export * from './users/admin.model';
-export * from './users/sponsor.model';
+
+export * from './fsae-user.model';
+export * from './roles';
+export * from './member.model';
+export * from './admin.model';
+export * from './sponsor.model';
 export * from './alumni.model';
